requirements.txt
langchain==0.0.123
openai==1.10.0
tree-sitter>=0.20.1
tiktoken>=0.3.0
jinja2>=3.1.2
datasets>=2.10.1
pandas>=1.5.3
unstructured>=0.5.7
numpy>=1.23.5
scikit-learn>=1.2.2
transformers>=4.27.4
torch>=2.0.0
gitpython>=3.1.31
unidiff>=0.7.5
tqdm>=4.65.0
python-dotenv
mypy
bandit